backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-cors-headers==4.1.0
django-filter==23.2
django-js-asset==2.1.0
django-mptt==0.14.0
djangorestframework==3.14.0
docopt==0.6.2
Faker==19.1.0
idna==3.4
Pillow==10.0.0
pipreqs==0.4.13
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.4
yarg==0.1.9
